% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_DEGs_DESeq2.R
\name{call_DEGs_DESeq2}
\alias{call_DEGs_DESeq2}
\title{Identifying Differentially Expressed Genes Using DESeq2}
\usage{
call_DEGs_DESeq2(data, sample, group, log2FoldChange = 1, padj = 0.05)
}
\arguments{
\item{data}{Expression matrix must be integers, with rows representing gene names and columns representing sample names.}

\item{sample}{Sample information table, with rows representing sample names.}

\item{group}{Grouping information of samples, the column names in the @param sample table, must be factors.}

\item{log2FoldChange}{The threshold for the absolute value of log2FoldChange is set to 1 by default.}

\item{padj}{The threshold for the adjusted p-value is set to 0.05 by default.}
}
\value{
A data frame.
}
\description{
Identifying Differentially Expressed Genes Using DESeq2
}
\examples{

library(dplyr)
library(biohelpers)

data(df.gene.rnaseq)
data(df.sample.rnaseq)
call_DEGs_DESeq2(data = df.gene.rnaseq, 
                 sample = df.sample.rnaseq, 
                 group = "group") -> degs

}
