% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommonlyUsed.anova_and_post_test.R
\name{CommonlyUsed.anova_and_post_test}
\alias{CommonlyUsed.anova_and_post_test}
\title{Title}
\usage{
CommonlyUsed.anova_and_post_test(
  data,
  group,
  value,
  method = "Tukey",
  level = 0.95
)
}
\arguments{
\item{data}{A data frame.}

\item{group}{The column of group information.}

\item{value}{The column of value.}

\item{method}{The default method is `Tukey`.}

\item{level}{The default level value is 0.95.}
}
\description{
Title
}
\examples{

library(dplyr)
library(biohelpers)

data(iris)

CommonlyUsed.anova_and_post_test(iris, "Species", "Sepal.Length") -> anova.result

}
