% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommonlyUsed.df_to_list.R
\name{CommonlyUsed.df_to_list}
\alias{CommonlyUsed.df_to_list}
\title{Convert data frame to list for venn plot.}
\usage{
CommonlyUsed.df_to_list(data, group, value)
}
\arguments{
\item{data}{A data frame.}

\item{group}{The column name for group.}

\item{value}{The column name for value.}
}
\value{
A list.
}
\description{
Convert data frame to list for venn plot.
}
\examples{

library(dplyr)
library(biohelpers)

data.frame(group = "A", value = 1:10) -> df.1
data.frame(group = "B", value = 3:12) -> df.2
data.frame(group = "C", value = 7:20) -> df.3

dplyr::bind_rows(df.1, df.2, df.3) \%>\%
  CommonlyUsed.df_to_list(group = "group", value = "value") -> res.list

}
