% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommonlyUsed.find_outliner.R
\name{CommonlyUsed.find_outliner}
\alias{CommonlyUsed.find_outliner}
\title{Identification of Outliers.}
\usage{
CommonlyUsed.find_outliner(x)
}
\arguments{
\item{x}{Numeric Format Vector.}
}
\value{
Vector. If it is an outlier, return 'yes'; otherwise, return 'no'.
}
\description{
Identification of Outliers.
}
\examples{

library(dplyr)
library(biohelpers)

data(iris)

iris \%>\%
  dplyr::group_by(Species) \%>\%
  dplyr::mutate(is.out = CommonlyUsed.find_outliner(Sepal.Length)) \%>\%
  dplyr::ungroup() -> result.find_outlinr

}
