% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_and_plot.R
\name{cor_and_plot}
\alias{cor_and_plot}
\title{Calculate Correlation, Return Correlation Results.}
\usage{
cor_and_plot(
  data.1,
  data.2 = NULL,
  method = "pearson",
  cor = 0.05,
  pvalue = 0.05
)
}
\arguments{
\item{data.1}{First Data Frame. Each column represents a feature value.}

\item{data.2}{Second Data Frame. Each column represents a feature value.}

\item{method}{There are three options for calculating correlation: `pearson`, `kendall`,and `spearman`, with `pearson` being the default.}

\item{cor}{Correlation threshold, default is 0.5.}

\item{pvalue}{P-value threshold, default is 0.05.}
}
\value{
A data frame containing the calculation results.
}
\description{
Calculate Correlation, Return Correlation Results.
}
\examples{

library(biohelpers)

x <- iris[, 1:4]
y <- iris[, 1:4]

cor_and_plot(data.1 = x, data.2 = y) -> result.cor

}
