% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.pcoa.sample}
\alias{df.pcoa.sample}
\title{Test data.}
\format{
## `df.pcoa.sample`
A data frame with 9 rows and 2 columns:
\describe{
  \item{sample}{sample name}
  \item{group}{sample group}
}
}
\source{
Xiang Li
}
\usage{
df.pcoa.sample
}
\description{
Test data.
}
\keyword{datasets}
