% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.permanova.sample}
\alias{df.permanova.sample}
\title{Test data.}
\format{
## `df.permanova.sample`
A data frame with 9 rows and 2 columns:
\describe{
}
}
\source{
Xiang Li
}
\usage{
df.permanova.sample
}
\description{
Test data.
}
\keyword{datasets}
