% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.rda.otu}
\alias{df.rda.otu}
\title{Test data.}
\format{
## `df.rda.otu`
A data frame with 9 rows and 2 columns:
\describe{
}
}
\source{
Xiang Li
}
\usage{
df.rda.otu
}
\description{
Test data.
}
\keyword{datasets}
