% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.reorder2heatmap}
\alias{df.reorder2heatmap}
\title{Test data.}
\format{
## `df.reorder2heatmap`
A data frame with 4,270 rows and 3 columns:
\describe{
  \item{sample}{sample code}
  \item{meta}{metabolite name}
  \item{value}{metabolite abundance}
}
}
\source{
Xiang Li
}
\usage{
df.reorder2heatmap
}
\description{
Test data.
}
\keyword{datasets}
