% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.rnaseq.degs}
\alias{df.rnaseq.degs}
\title{Test data.}
\format{
## `df.rnaseq.degs`
A data frame with 6 rows and 1 columns:
\describe{
  \item{gene}{gene name}
}
}
\source{
Xiang Li
}
\usage{
df.rnaseq.degs
}
\description{
Test data.
}
\keyword{datasets}
