% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.rnaseq.gene}
\alias{df.rnaseq.gene}
\title{Test data.}
\format{
## `df.rnaseq.gene`
A data frame with 57,359 rows and 6 columns:
\describe{
  \item{SRR12580270}{sample name}
  \item{SRR12580269}{sample name}
  \item{SRR12580268}{sample name}
  \item{SRR12580258}{sample name}
  \item{SRR12580257}{sample name}
  \item{SRR12580256}{sample name}
}
}
\source{
Xiang Li
}
\usage{
df.rnaseq.gene
}
\description{
Test data.
}
\keyword{datasets}
