% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.rnaseq.kegg}
\alias{df.rnaseq.kegg}
\title{Test data.}
\format{
## `df.rnaseq.kegg`
A data frame with 6 rows and 1 columns:
\describe{
  \item{gene}{gene name}
  \item{kegg.id}{KEGG term id}
  \item{kegg.term}{KEGG term name}
}
}
\source{
Xiang Li
}
\usage{
df.rnaseq.kegg
}
\description{
Test data.
}
\keyword{datasets}
