% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.rnaseq.plot_volcano}
\alias{df.rnaseq.plot_volcano}
\title{Test data.}
\format{
## `df.rnaseq.plot_volcano`
A data frame with 57,359 rows and 8 columns:
\describe{
  \item{gene}{gene name}
  \item{baseMea}{baseMea}
  \item{log2FoldChange}{log2FoldChange}
  \item{lfcSE}{lfcSE}
  \item{stat}{stat}
  \item{pvalue}{pvalue}
  \item{padj}{padj}
  \item{group}{group}
}
}
\source{
Xiang Li
}
\usage{
df.rnaseq.plot_volcano
}
\description{
Test data.
}
\keyword{datasets}
