% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.rnaseq.sample}
\alias{df.rnaseq.sample}
\title{Test data.}
\format{
## `df.rnaseq.sample`
A data frame with 6 rows and 1 columns:
\describe{
  \item{group}{sample group}
}
}
\source{
Xiang Li
}
\usage{
df.rnaseq.sample
}
\description{
Test data.
}
\keyword{datasets}
