% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.splsda.meta}
\alias{df.splsda.meta}
\title{Test data.}
\format{
## `df.splsda.meta`
A data frame with 9 rows and 2 columns:
\describe{
}
}
\source{
Xiang Li
}
\usage{
df.splsda.meta
}
\description{
Test data.
}
\keyword{datasets}
