% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.top10.class}
\alias{df.top10.class}
\title{Test data.}
\format{
## `df.top10.calss`
A data frame with 9 rows and 2 columns:
\describe{
}
}
\source{
Xiang Li
}
\usage{
df.top10.class
}
\description{
Test data.
}
\keyword{datasets}
