% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metabolome.oplsda.R
\name{Metabolome.oplsda}
\alias{Metabolome.oplsda}
\title{OPLS-DA.}
\usage{
Metabolome.oplsda(data, group, VIP = 1)
}
\arguments{
\item{data}{Numerical matrix, (observations x variables; NAs are allowed), data.frame, SummarizedExperiment or ExpressionSet object.}

\item{group}{A factor (same length as 'x' row number) for (O)PLS-DA}

\item{VIP}{Numerical vector of Variable Importance in Projection.}
}
\value{
A list.
}
\description{
OPLS-DA.
}
\examples{

library(dplyr)
library(biohelpers)

data("df.splsda.meta")
data("df.splsda.sample")

df.splsda.sample \%>\%
  dplyr::filter(day == 180, treatment != "Cover") \%>\%
  dplyr::mutate(treatment = factor(treatment)) -> df.oplsda.sample

df.splsda.meta \%>\%
  dplyr::filter(rownames(.) \%in\% df.oplsda.sample$sample) -> df.oplsda.meta

Metabolome.oplsda(df.oplsda.meta, df.oplsda.sample$treatment) -> oplsda.result

}
