% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metabolome.splsda.R
\name{Metabolome.splsda}
\alias{Metabolome.splsda}
\title{Sparse Partial Least Squares Discriminant Analysis (sPLS-DA)}
\usage{
Metabolome.splsda(data, group, ncomp = 3)
}
\arguments{
\item{data}{Numeric matrix of predictors with the rows as individual observations. missing values (NAs) are allowed.}

\item{group}{A factor or a class vector for the discrete outcome.}

\item{ncomp}{Positive Integer. The number of components to include in the model. Default to 3.}
}
\value{
A list.
}
\description{
Sparse Partial Least Squares Discriminant Analysis (sPLS-DA)
}
\examples{

library(tidyverse)
library(biohelpers)

data("df.splsda.meta")
data("df.splsda.sample")

Metabolome.splsda(df.splsda.meta, as.factor(df.splsda.sample$day)) -> splsda.result

}
