% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Microbiome.call_DAMs_DESeq2.R
\name{Microbiome.call_DAMs_DESeq2}
\alias{Microbiome.call_DAMs_DESeq2}
\title{Title}
\usage{
Microbiome.call_DAMs_DESeq2(
  data,
  sample,
  formula = ~group,
  log2FoldChange = 1,
  padj = 0.05
)
}
\arguments{
\item{data}{Expression matrix must be integers, with rows representing gene names and columns representing sample names.}

\item{sample}{Sample information table, with rows representing sample names.}

\item{formula}{A formula or matrix. the formula expresses how the counts for each gene depend on the variables in colData.}

\item{log2FoldChange}{A The threshold for the log2FoldChange.}

\item{padj}{The threshold for the adjusted p-value is set to 0.05 by default.}
}
\value{
A data frame from DESeq2.
}
\description{
Title
}
\examples{

library(biohelpers)

data(df.pcoa.otu)
data(df.pcoa.sample)

Microbiome.call_DAMs_DESeq2(t(df.pcoa.otu), df.pcoa.sample) -> dems.res

}
