% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Microbiome.call_DAMs_LEfSe.R
\name{Microbiome.call_DAMs_LEfSe}
\alias{Microbiome.call_DAMs_LEfSe}
\title{Identifying differential abundance microbiomes using LEfSe.}
\usage{
Microbiome.call_DAMs_LEfSe(
  data,
  sample,
  groupCol = "group",
  kruskal.threshold = 0.05,
  wilcox.threshold = 0.05,
  lda.threshold = 1
)
}
\arguments{
\item{data}{Feature table, with rows representing feature values and columns representing sample names.}

\item{sample}{Sample table, with rows representing sample name and columns representing group information.}

\item{groupCol}{The column name where the grouping information is located, default is 'group'.}

\item{kruskal.threshold}{The threshold for the Kruskal test, default is 0.05.}

\item{wilcox.threshold}{The threshold for the Wilcox test, default is 0.05.}

\item{lda.threshold}{The threshold for the LDA test, default is 1.}
}
\value{
A data frame.
}
\description{
Identifying differential abundance microbiomes using LEfSe.
}
\examples{

library(biohelpers)
library(dplyr)

data(df.call_DAMs_LEfSe.otu)
data(df.call_DAMs_LEfSe.sample)

Microbiome.call_DAMs_LEfSe(df.call_DAMs_LEfSe.otu, df.call_DAMs_LEfSe.sample) -> lefse.result

}
