% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Microbiome.permanova.R
\name{Microbiome.permanova}
\alias{Microbiome.permanova}
\title{Title}
\usage{
Microbiome.permanova(formula, sample, method = "bray", permutations = 999)
}
\arguments{
\item{formula}{A formula, such as df.permanova.otu ~ location * variety * group.}

\item{sample}{Sample information table.}

\item{method}{The name of any method used in vegdist to calculate pairwise distances if the left hand side of the formula was a data frame or a matrix.}

\item{permutations}{a list of control values for the permutations as returned by the function how, or the number of permutations required, or a permutation matrix where each row gives the permuted indices.}
}
\value{
A data frame.
}
\description{
Title
}
\examples{

library(dplyr)
library(biohelpers)

data(df.permanova.otu)
data(df.permanova.sample)

Microbiome.permanova(
  df.permanova.otu ~ location * variety * group,
  df.permanova.sample
) -> result.permanova

}
