% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Microbiome.RDA.R
\name{Microbiome.RDA}
\alias{Microbiome.RDA}
\title{Title}
\usage{
Microbiome.RDA(
  data,
  physicochemical,
  sample,
  method.decostand = "hellinger",
  scale = FALSE,
  color = "group"
)
}
\arguments{
\item{data}{A data frame where rows represent samples and columns represent OTUs.}

\item{physicochemical}{A physicochemical properties data frame where rows represent samples and columns represent physicochemical properties.}

\item{sample}{A sample information table where the first column is 'sample'.}

\item{method.decostand}{Method of vegan::decostand().}

\item{scale}{Scale or not in vegan::rda().}

\item{color}{Color for plot.}
}
\value{
A list contains a ggplot2  object and two data frames.
}
\description{
Title
}
\examples{

library(dplyr)
library(biohelpers)

data(df.rda.otu)
data(df.rda.chem)

df.rda.chem \%>\%
  rownames() \%>\%
  as.data.frame() \%>\%
  magrittr::set_names("sample") \%>\%
  dplyr::mutate(group = stringr::str_split(sample, "_") \%>\% sapply("[", 1)) -> df.sample

Microbiome.RDA(df.rda.otu, df.rda.chem, df.sample) -> rda.result

}
