% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Microbiome.rrarefy.R
\name{Microbiome.rrarefy}
\alias{Microbiome.rrarefy}
\title{Rarefaction Species Richness}
\usage{
Microbiome.rrarefy(data, value = NULL)
}
\arguments{
\item{data}{Community data, a matrix-like object or a vector.}

\item{value}{Subsample size for rarefying community, either a single value or a vector. The default is the minimum value in the sample.}
}
\value{
A data frame.
}
\description{
Rarefaction Species Richness
}
\examples{

library(dplyr)
library(biohelpers)

data(df.permanova.otu)

df.permanova.otu \%>\%
  t() \%>\%
  as.data.frame() \%>\%
  Microbiome.rrarefy() -> df.otu.rrarefy

}
