% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Microbiome.top_10.R
\name{Microbiome.top_10}
\alias{Microbiome.top_10}
\title{Calculate the top ten microorganisms by abundance at different taxonomic levels.}
\usage{
Microbiome.top_10(
  data,
  sample,
  taxo,
  which = "phylum",
  by = "sum",
  method = "wilcox",
  ref = "CK",
  pvalue = 0.05
)
}
\arguments{
\item{data}{Feature table, such as OTU abundance information, where rows represent samples and columns represent feature values.}

\item{sample}{Sample table, where the first column is the sample name, the second column and onwards contain grouping information, and there must be a column named 'group'.}

\item{taxo}{Classification information, such as the classification information of OTUs.}

\item{which}{Which classification information to analyze and plot, with 'phylum' as the default.}

\item{by}{Calculate the Top 10 by summation or averaging, with 'sum' as the default.}

\item{method}{Statistical analysis method, with 'wilcox' (Wilcox test) as the default.}

\item{ref}{The name of the control group used for statistical analysis, with 'CK' as the default.}

\item{pvalue}{Threshold for the p-value, with a default of 0.05.}
}
\value{
A list containing a data frame and a ggplot2 object.
}
\description{
Calculate the top ten microorganisms by abundance at different taxonomic levels.
}
\examples{

library(dplyr)
library(biohelpers)

data(df.top10.otu)
data(df.top10.sample)
data(df.top10.class)

Microbiome.top_10(df.top10.otu, df.top10.sample, df.top10.class) -> top10.res

}
