% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAinALL.R
\name{PCAinALL}
\alias{PCAinALL}
\title{Perform principal component analysis (PCA) and return the results.}
\usage{
PCAinALL(data, sample, pca.num = 5, plot = TRUE)
}
\arguments{
\item{data}{A data frame where rows are samples and columns are indicators.}

\item{sample}{A data frame containing sample information, where one column is the sample names that match those in {\data}, without the need for row names.}

\item{pca.num}{The returned results include the number of principal components, with a default value of 10.}

\item{plot}{Whether to plot, with the default being to draw a scatter plot. The plotted graph is a ggplot2 object, which can be customized and enhanced using ggplot2.}
}
\description{
Perform principal component analysis (PCA) and return the results.
}
