% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopGen.manhattan_plot.R
\name{PopGen.manhattan_plot}
\alias{PopGen.manhattan_plot}
\title{Plot the manhattan from a GWAS results.}
\usage{
PopGen.manhattan_plot(df, p.threshold = NA)
}
\arguments{
\item{df}{A data frame.}

\item{p.threshold}{The threshold of P value, the defalut value is 0.05 / number of SNPs.}
}
\value{
A ggplot2 object.
}
\description{
Plot the manhattan from a GWAS results.
}
\examples{

library(tidyverse)

# 设置随机数种子
set.seed(20250312)

# 构造GWAS数据
normentR::simulateGWAS(nSNPs = 1e6, nSigCols = 3) \%>\%
  janitor::clean_names() \%>\%
  dplyr::rename(posi = bp) \%>\%
  dplyr::mutate(label = case_when(p < 0.05 / nrow(.) ~ snp, TRUE ~ NA)) -> df.gwas

PopGen.manhattan_plot(df.gwas)

}
