% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeq.call_DEGs_DESeq2.R
\name{RNASeq.call_DEGs_DESeq2}
\alias{RNASeq.call_DEGs_DESeq2}
\title{Identifying Differentially Expressed Genes Using DESeq2}
\usage{
RNASeq.call_DEGs_DESeq2(
  data,
  sample,
  formula = ~group,
  log2FoldChange = 1,
  padj = 0.05
)
}
\arguments{
\item{data}{Expression matrix must be integers, with rows representing gene names and columns representing sample names.}

\item{sample}{Sample information table, with rows representing sample names.}

\item{log2FoldChange}{The threshold for the absolute value of log2FoldChange is set to 1 by default.}

\item{padj}{The threshold for the adjusted p-value is set to 0.05 by default.}
}
\value{
A data frame.
}
\description{
Identifying Differentially Expressed Genes Using DESeq2
}
\examples{

library(dplyr)
library(biohelpers)

data(df.rnaseq.gene)
data(df.rnaseq.sample)
RNASeq.call_DEGs_DESeq2(
  data = df.rnaseq.gene,
  sample = df.rnaseq.sample
) -> degs

}
