% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeq.enrich_KEGG.R
\name{RNASeq.enrich_KEGG}
\alias{RNASeq.enrich_KEGG}
\title{Perform KEGG enrichment analysis of differentially expressed genes.}
\usage{
RNASeq.enrich_KEGG(gene, kegg.db, pAdjustMethod = "BH", p.adjust = 0.05)
}
\arguments{
\item{gene}{The names of differentially expressed genes are a vector.}

\item{kegg.db}{The KEGG database used for KEGG enrichment analysis needs to
be in data frame format and must contain the following columns: gene, kegg.id, kegg.term.
The kegg.ontology column is optional.}

\item{pAdjustMethod}{Methods for adjusting p-values can be selected from "holm", "hochberg",
"hommel", "bonferroni", "BH", "BY", "fdr", and "none", with "BH" as the default.}

\item{p.adjust}{Threshold for the adjusted P value.}
}
\value{
A data frame.
}
\description{
Perform KEGG enrichment analysis of differentially expressed genes.
}
\examples{

library(biohelpers)

data(df.rnaseq.kegg)
data(df.rnaseq.degs)

RNASeq.enrich_KEGG(gene = df.rnaseq.degs$gene, kegg.db = df.rnaseq.kegg) -> kegg.res

}
