% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Test data.}
\format{
## `test_data`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{Sepal.Length}{length of sepal}
  \item{Sepal.Width}{width of sepal}
  \item{Petal.Length}{length of petal}
  \item{Petal.Width}{width of petal}
  \item{Species}{species group}
}
}
\source{
<https://archive.ics.uci.edu/dataset/53/iris>
}
\usage{
test_data
}
\description{
Test data.
}
\keyword{datasets}
