% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biohelpers-package.R
\docType{package}
\name{biohelpers-package}
\alias{biohelpers}
\alias{biohelpers-package}
\title{biohelpers: Convenience Functions for Biological Data Processing}
\description{
Convenience Functions for Biological Data Processing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lixiang117423/biohelpers}
  \item \url{https://lixiang117423.github.io/biohelpers/}
  \item Report bugs at \url{https://github.com/lixiang117423/biohelpers/issues}
}

}
\author{
\strong{Maintainer}: Xiang LI \email{lixiang117423@gmail.com}

}
\keyword{internal}
