% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_GO.R
\name{enrich_GO}
\alias{enrich_GO}
\title{Perform GO enrichment analysis of differentially expressed genes.}
\usage{
enrich_GO(gene, go.db, pAdjustMethod = "BH", p.adjust = 0.05)
}
\arguments{
\item{gene}{The names of differentially expressed genes are a vector.}

\item{go.db}{The GO database used for GO enrichment analysis needs to
be in data frame format and must contain the following columns: gene, go.id, go.term.
The go.ontology column is optional.}

\item{pAdjustMethod}{Methods for adjusting p-values can be selected from "holm", "hochberg",
"hommel", "bonferroni", "BH", "BY", "fdr", and "none", with "BH" as the default.}

\item{p.adjust}{Threshold for the adjusted P value.}
}
\value{
A data frame.
}
\description{
Perform GO enrichment analysis of differentially expressed genes.
}
\examples{

library(biohelpers)

data(df.rnaseq.go)
data(df.rnaseq.degs)

enrich_GO(gene = df.rnaseq.degs$gene, go.db = df.rnaseq.go) -> go.res

}
