% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Microbiome.call_DAMs_DESeq2.R
\name{Microbiome.call_DAMs_DESeq2}
\alias{Microbiome.call_DAMs_DESeq2}
\title{Title}
\usage{
Microbiome.call_DAMs_DESeq2(
  data,
  sample,
  group = "group",
  log2FoldChange = 1,
  padj = 0.05
)
}
\arguments{
\item{data}{Expression matrix must be integers, with rows representing gene names and columns representing sample names.}

\item{sample}{Sample information table, with rows representing sample names.}

\item{group}{Grouping information of samples, the column names in the @param sample table, must be factors.}

\item{log2FoldChange}{The threshold for the absolute value of log2FoldChange is set to 1 by default.}

\item{padj}{The threshold for the adjusted p-value is set to 0.05 by default.}
}
\value{
A data frame from DESeq2.
}
\description{
Title
}
\examples{

library(biohelpers)

data(df.pcoa.otu)
data(df.pcoa.sample)

Microbiome.call_DAMs_DESeq2(t(df.pcoa.otu), df.pcoa.sample) -> dems.res

}
