% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Create a volcano plot.}
\usage{
plot_volcano(
  data,
  x = "log2FoldChange",
  y = "padj",
  size = "baseMean",
  color = "group",
  xintercept.abs = 1,
  yintercept = 0.05
)
}
\arguments{
\item{data}{The data frame used to create the scatter plot is typically the output of DESeq2, specifically the output from call_DEGs_DESeq2.}

\item{size}{Column name used to map the size of the points, default is baseMean.}

\item{color}{Column name used to map the color of the points.}

\item{xintercept.abs}{Intercept of X axes. It is usually log2FoldChange, which includes both positive and negative values.}

\item{yintercept}{Intercept of Y axes. It is usually -log10(padj).}
}
\value{
A ggplot2 object.
}
\description{
Create a volcano plot.
}
\examples{

library(biohelpers)

data(df.rnaseq.plot_volcano)

plot_volcano(data = df.rnaseq.plot_volcano) -> p

}
